changed int to float

+++ main.py
@@ -1,3 +1,3 @@
-temp = int(input('Enter temperature in celsius: '))
+temp = float(input('Enter temperature in celsius: '))
 m = (temp * (9/5)) + 32
 print(str(temp) + '° in Celsius is equivalent to ' + str(m) + '° fahrenheit ')
